requirements.txt
altgraph==0.17.4
biopython==1.83
chardet==5.2.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
macholib==1.16.3
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyMuPDF==1.23.18
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.9
six==1.16.0
tzdata==2023.4
